+++ src/cn/langya/sun/values/ColorValue.java
@@ -30,13 +30,13 @@ public class ColorValue extends AbstractValue<Integer>
             return new float[] { 0.0f, 0.0f, 0.0f };
         }
         final float[] hsbValues = new float[3];
-        int cMax = Math.max((int)color >>> 16 & 0xFF, (int)color >>> 8 & 0xFF);
-        if (((int)color & 0xFF) > cMax) {
-            cMax = ((int)color & 0xFF);
+        int cMax = Math.max(color >>> 16 & 0xFF, color >>> 8 & 0xFF);
+        if ((color & 0xFF) > cMax) {
+            cMax = (color & 0xFF);
         }
-        int cMin = Math.min((int)color >>> 16 & 0xFF, (int)color >>> 8 & 0xFF);
-        if (((int)color & 0xFF) < cMin) {
-            cMin = ((int)color & 0xFF);
+        int cMin = Math.min(color >>> 16 & 0xFF, color >>> 8 & 0xFF);
+        if ((color & 0xFF) < cMin) {
+            cMin = (color & 0xFF);
         }
         final float brightness = cMax / 255.0f;
         final float saturation = (cMax != 0) ? ((cMax - cMin) / (float)cMax) : 0.0f;
@@ -45,10 +45,10 @@ public class ColorValue extends AbstractValue<Integer>
             hue = 0.0f;
         }
         else {
-            final float redC = (cMax - ((int)color >>> 16 & 0xFF)) / (float)(cMax - cMin);
-            final float greenC = (cMax - ((int)color >>> 8 & 0xFF)) / (float)(cMax - cMin);
-            final float blueC = (cMax - ((int)color & 0xFF)) / (float)(cMax - cMin);
-            hue = ((((int)color >>> 16 & 0xFF) == cMax) ? (blueC - greenC) : ((((int)color >>> 8 & 0xFF) == cMax) ? (2.0f + redC - blueC) : (4.0f + greenC - redC))) / 6.0f;
+            final float redC = (cMax - (color >>> 16 & 0xFF)) / (float)(cMax - cMin);
+            final float greenC = (cMax - (color >>> 8 & 0xFF)) / (float)(cMax - cMin);
+            final float blueC = (cMax - (color & 0xFF)) / (float)(cMax - cMin);
+            hue = (((color >>> 16 & 0xFF) == cMax) ? (blueC - greenC) : (((color >>> 8 & 0xFF) == cMax) ? (2.0f + redC - blueC) : (4.0f + greenC - redC))) / 6.0f;
             if (hue < 0.0f) {
                 ++hue;
             }
