+++ src/net/minecraft/entity/ai/EntityAIHarvestFarmland.java
@@ -60,7 +60,7 @@ public class EntityAIHarvestFarmland extends EntityAIMoveToBlock
     public void updateTask()
     {
         super.updateTask();
-        this.theVillager.getLookHelper().setLookPosition((double)this.destinationBlock.getX() + 0.5D, (double)(this.destinationBlock.getY() + 1), (double)this.destinationBlock.getZ() + 0.5D, 10.0F, (float)this.theVillager.getVerticalFaceSpeed());
+        this.theVillager.getLookHelper().setLookPosition((double)this.destinationBlock.getX() + 0.5D, this.destinationBlock.getY() + 1, (double)this.destinationBlock.getZ() + 0.5D, 10.0F, (float)this.theVillager.getVerticalFaceSpeed());
 
         if (this.getIsAboveDestination())
         {
